requirements.txt
Django>=4.2,<4.5
wagtail>=5.0.2,<5.3
psycopg2==2.9.2
django-compressor
wagtail>=5.1,<5.2
django_extensions>=3.2.3
django-widget-tweaks>=1.5.0
sentry-sdk>=1.31.0
django-debug-toolbar>=4.2.0,<=4.9.0
whitenoise>=6.5.0,<7.0.0
django-allauth
python-dateutil
openpyxl
black
crispy-bootstrap5==0.7
django-crispy-forms==2.0